Spring Boot Application with Mutable Health Indicator
This change is the initial commit of a Spring Boot application that
demonstrates integration with the PivotalCF Applications Manager.  This
integration consists of exposure of the Info and Health Actuator endpoint
payloads in a user-pleasing UI.

In addition to this support, the application exposes a REST endpoint that
allows the caller to adjust a made-for-purpose HealthIndicator which should
then affect the health of the entire application.  To use this endpoint, POST
to /mutate with an application/x-www-form-urlencoded payload containing a
required status parameter with values matching the
org.springframework.boot.actuate.health.Status enumeration options, as well as
optional detailKey and detailValue Strings that will be added to the health
output.