Composite Mutable Health Indicator
Previously, the MutableHealthIndicator only exposed a top-level
HealthIndicator for mutation.  As part of testing with other applications, the
fact there there were multi-level nested CompositeHealthIndicators came to
light.  This change adds a MutableCompositeHealthIndicator which exposes this
multi-level nested behavior.

In addition to the exposure of a new type and mutation endpoints, this change
also updates the payload available to mutation.  In addition to the previous
status RequestParam, the entire body of the request, in JSON form, will be
added to the mutated HealthIndicator.  This change means that very complex
payloads, (e.g. maps containing arrays) can be simulated for testing.