Disable CF SSL Validation
Previously, the application used the default configuration for the Apps
Manager Spring Boot integration.  Unfortunately, many of the places the
application is pushed use self-signed certificates.  This change disables SSL
validation when contacting the Cloud Controller to ensure that the application
runs in these situations.

DO NOT use this as a pattern in real applications.